trademark_checker/requirements.txt
streamlit>=1.22.0
requests>=2.31.0
fpdf2>=2.8.0
pandas>=2.0.0
